Improvement: minor changes to the makefile to have better output and a quicker buildtools target.

+++ src/autoscaler/api/cmd/api/api_suite_test.go
@@ -204,8 +204,12 @@ var _ = SynchronizedBeforeSuite(func() []byte {
 })
 
 var _ = SynchronizedAfterSuite(func() {
-	os.Remove(configFile.Name())
-	ccServer.Close()
+	if configFile != nil {
+		_ = os.Remove(configFile.Name())
+	}
+	if ccServer != nil {
+		ccServer.Close()
+	}
 }, func() {
 	gexec.CleanupBuildArtifacts()
 })
